pyproject.toml
[project]
name = "memory_engine"
version = "0.0.0"
description = "Reference implementation of the Wiki v3 neural memory orchestration blueprint."
readme = "README.md"
requires-python = ">=3.12"
license = { file = "LICENSE" }
authors = [{ name = "randunun-eng" }]
keywords = ["memory", "digital-twin", "llm", "rag", "orchestration"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "aiosqlite>=0.20.0",
    "sqlite-vec>=0.1.6",
    "pynacl>=1.5.0",
    "pydantic>=2.7.0",
    "pydantic-settings>=2.3.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "click>=8.1.0",
    "orjson>=3.10.0",
    "httpx>=0.27.0",
    "rank-bm25>=0.2.2",
    "sentence-transformers>=3.0.0",
    "PyYAML>=6.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=5.0.0",
    "pytest-xdist>=3.6.0",
    "hypothesis>=6.110.0",
    "ruff>=0.6.0",
    "mypy>=1.11.0",
    "pre-commit>=4.0.0",
    "types-click",
    "types-PyYAML",
]
postgres = [
    "asyncpg>=0.29.0",
    "pgvector>=0.3.0",
]
all = [
    "memory_engine[dev,postgres]",
]

[project.scripts]
memory-engine = "memory_engine.cli.main:main"

[project.urls]
Homepage = "https://github.com/randunun-eng/memory_engine"
Documentation = "https://github.com/randunun-eng/memory_engine/blob/main/CLAUDE.md"
Repository = "https://github.com/randunun-eng/memory_engine"
Issues = "https://github.com/randunun-eng/memory_engine/issues"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/memory_engine"]

# ----------------------------------------------------------------------
# Ruff
# ----------------------------------------------------------------------

[tool.ruff]
line-length = 100
target-version = "py312"
src = ["src", "tests"]
# twincore-alpha/ is the reference deployment (twin-agent, control-plane,
# whatsapp-bridge, persona YAMLs, bootstrap scripts). It's tracked in this
# repo for traceability but isn't part of the memory_engine library; it
# has different style conventions (Greek α in milestone labels, its own
# pyproject where relevant). Excluding from ruff lint + format keeps CI
# focused on the library surface. See twincore-alpha/DRIFT.md for its
# own conventions.
extend-exclude = ["twincore-alpha"]

[tool.ruff.lint]
select = [
    "E",     # pycodestyle errors
    "W",     # pycodestyle warnings
    "F",     # pyflakes
    "I",     # isort
    "B",     # flake8-bugbear
    "C4",    # flake8-comprehensions
    "S",     # flake8-bandit (security)
    "UP",    # pyupgrade
    "ASYNC", # flake8-async
    "RUF",   # ruff-specific
    "SIM",   # flake8-simplify
    "TCH",   # flake8-type-checking
    "N",     # pep8-naming
]
ignore = [
    "E501",  # line-too-long (formatter handles)
    "B008",  # function calls in argument defaults (Pydantic/Click pattern)
    "S101",  # use of assert (fine in tests, filtered per-file)
]

[tool.ruff.lint.per-file-ignores]
"tests/**" = [
    "S101",   # asserts allowed
    "S105",   # hardcoded passwords allowed
    "S106",   # hardcoded passwords allowed
    "S311",   # insecure random allowed (test fixtures)
]
"examples/**" = ["S311", "T20"]  # allow print and insecure random in examples
"src/memory_engine/exceptions.py" = ["N818"]  # exception names match blueprint spec
"src/memory_engine/retrieval/bm25.py" = ["S608"]  # SQL uses parameterized LensFilter, safe
"src/memory_engine/retrieval/vector.py" = ["S608"]  # SQL uses parameterized LensFilter, safe
"src/memory_engine/healing/invariants.py" = ["S608"]  # WHERE clauses built from hardcoded strings, not user input
"tests/invariants/test_phase1_retrieval.py" = ["N802"]  # T3 canary test names per spec
"tests/invariants/test_phase5.py" = ["N802"]  # T3/T11 release gate test names per spec
"tests/integration/test_phase6.py" = ["S603", "S607"]  # subprocess calls to our own bin/ scripts

[tool.ruff.lint.isort]
known-first-party = ["memory_engine"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "lf"

# ----------------------------------------------------------------------
# mypy
# ----------------------------------------------------------------------

[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_configs = true
show_error_codes = true
pretty = true

[[tool.mypy.overrides]]
module = [
    "nacl.*",
    "sqlite_vec.*",
    "sentence_transformers.*",
    "rank_bm25.*",
]
ignore_missing_imports = true

# ----------------------------------------------------------------------
# pytest
# ----------------------------------------------------------------------

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = [
    "--strict-markers",
    "-ra",
    "--tb=short",
]
markers = [
    "eval: slow evaluation tests, excluded from default run",
    "perf: performance benchmarks (10k+ neurons), excluded from default run",
    "postgres: requires a running Postgres instance",
]

# ----------------------------------------------------------------------
# coverage
# ----------------------------------------------------------------------

[tool.coverage.run]
source = ["src/memory_engine"]
branch = true
omit = [
    "*/tests/*",
    "*/__init__.py",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
    "if __name__ == .__main__.:",
]
show_missing = true
